fixed quoting

+++ src/main/java/RFQRunnable.java
@@ -43,14 +43,14 @@ public class RFQRunnable implements Runnable {
 
         public double sendRequest(int buyOrSell, double quantity) {
             if (buyOrSell == 1) {
-                if (oBook.getBuyPrices().get(Math.floor(quantity / 100) * 100) != null) {
-                    return oBook.getBuyPrices().get(Math.floor(quantity / 100) * 100);
+                if (oBook.getBuyPrices().get((int)Math.floor(quantity / 100) * 100) != null) {
+                    return oBook.getBuyPrices().get((int)Math.floor(quantity / 100) * 100);
                 } else {
                     return -1;
                 }
             } else if (buyOrSell == -1) {
-                if (oBook.getSellPrices().get(Math.ceil(quantity / 100) * 100) != null) {
-                    return oBook.getSellPrices().get(Math.ceil(quantity / 100) * 100);
+                if (oBook.getSellPrices().get((int)Math.ceil(quantity / 100) * 100) != null) {
+                    return oBook.getSellPrices().get((int)Math.ceil(quantity / 100) * 100);
                 } else {
                     return -1;
                 }
